(feat): Performance tweaks
Inject a pre-configured wisski_py.Api instance, rather than constructing
one each time `DocumentEntity` is constructed.

Cache calls that retrieve entity-URIs via `functools.cache`.
This decorator caches return values, and uses the wrapped function's
arguments as the key; thus, the arguments all need to be hashable. For
this reason, we introduce NamedTuples that hold the expected
format-string variables.

Refactor `FieldFunctions` into a class that is initialized once for each
`DocumentEntity`. `FieldFunctions.exception(field_name)` returns a
function that can either be passed as a callable, or called directly
with further arguments.
This way, we only instantiate `FieldFunctions` once for a document, and
not once for each call to the exception helper.

Bump the wisski_py submodule to the most recent version, which contains
a feature for lazily loading pathbuilders into the API.