(fix): Remove functools' LRU cache
The functools cache was aggressively caching any return values from
`entity_uri`, including `None`-values. However, if a `None` is
encountered, the missing entity is created. With functools, the cache is
never updated, causing the `None` to persist throughout and leading to
data inconsistency.

This patch introduces a local dict acting as a cache, which is updated
whenever an entity is found, even after subsequent insertion.
The keys to this cache are the "search value", i.e., the name of the
thing that is being queried, and a hash of the SPARQL query. This should
avoid key collisions.